1 meetplayer fix listview position problem 2 engine2 modify osles setVol()

+++ engine2/platform/oslesrender.h
@@ -20,7 +20,7 @@ public:
 	
 	void flush();
 
-	int setVol(int vol); // 0-100
+	int setVol(double vol); // 0-1
 
 	int getVol();
